refactor: expose llm instance and drop wrappers
